+++ solr/solrj/src/test/org/apache/solr/client/solrj/response/QueryResponseTest.java
@@ -130,7 +130,7 @@ public class QueryResponseTest extends SolrTestCase {
     GroupCommand fieldCommand = commands.get(0);
     assertEquals("acco_id", fieldCommand.getName());
     assertEquals(30000000, fieldCommand.getMatches());
-    assertEquals(5687, fieldCommand.getNGroups().intValue());
+    assertEquals(5687l, fieldCommand.getNGroups().longValue());
     List<Group> fieldCommandGroups = fieldCommand.getValues();
     assertEquals(10, fieldCommandGroups.size());
     assertEquals("116_ar", fieldCommandGroups.get(0).getGroupValue());
@@ -236,7 +236,7 @@ public class QueryResponseTest extends SolrTestCase {
     GroupCommand fieldCommand = commands.get(0);
     assertEquals("acco_id", fieldCommand.getName());
     assertEquals(30000000, fieldCommand.getMatches());
-    assertEquals(5687, fieldCommand.getNGroups().intValue());
+    assertEquals(5687l, fieldCommand.getNGroups().longValue());
     List<Group> fieldCommandGroups = fieldCommand.getValues();
     assertEquals(1, fieldCommandGroups.size());
